Move dialogs to separate file
