+++ fv1_programmer/tui.css
@@ -112,76 +112,6 @@ MarkdownFence {
     offset-y: 100%;
 }
 
-ModalScreen {
-    align: center middle;
-}
-
-#yesnodialog {
-    grid-size: 2;
-    grid-gutter: 1 2;
-    grid-rows: 1fr 3;
-    padding: 0 1;
-    width: 60;
-    height: 11;
-    border: thick $background 80%;
-    background: $surface;
-}
-
-#yesnomessage {
-    column-span: 2;
-    height: 1fr;
-    width: 1fr;
-    content-align: center middle;
-}
-
-#yesnodialog > Button {
-    width: 100%;
-}
-
-#filedialog {
-    grid-size: 2;
-    grid-gutter: 1 2;
-    grid-rows: 1fr 3;
-    padding: 0 1;
-    width: 90vw;
-    height: 90vh;
-    border: thick $background 80%;
-    background: $surface;
-}
-
-#renameslotdialog {
-    padding: 0 1;
-    width: 80vw;
-    height: 30vh;
-    border: thick $background 80%;
-    background: $surface;
-}
-
-#renameslotdialog > Input {
-    margin: 1 1;
-    border: $primary;
-}
-
-#filesavedialog {
-    padding: 0 1;
-    width: 80vw;
-    height: 30vh;
-    border: thick $background 80%;
-    background: $surface;
-}
-
-#filesavedialog > Input {
-    margin: 1 1;
-    border: $primary;
-}
-
-#fileselectlabel {
-    column-span: 2;
-    content-align: left bottom;
-    width: 100%;
-    height: auto;
-    dock: top;
-}
 
 #filetree {
     column-span: 2;
@@ -190,36 +120,4 @@ ModalScreen {
     height: 100%;
     margin-top: 1;
     margin-bottom: 1;
-}
-
-#filedialogcancel {
-    dock: bottom;
-    width: 30;
-    margin-left: 5;
-    align: left top;
-}
-
-#filedialogselect {
-    dock: bottom;
-    width: 30;
-    margin-right: 5;
-    align: right top;
-}
-
-#busyscreen {
-    grid-size: 1;
-    grid-gutter: 1 1;
-    grid-rows: 1fr 4;
-    padding: 0 1;
-    width: 40;
-    height: 10;
-    border: thick $background 80%;
-    background: $surface;
-}
-
-#busyscreen > Label {
-    height: 1fr;
-}
-#busyscreen > LoadingIndicator {
-    height: 3fr;
 }
